--- src/main/java/com/ludo/study/studymatchingplatform/auth/kakao/service/dto/response/AuthenticationResponse.java
@@ -1,8 +1,0 @@
-package com.ludo.study.studymatchingplatform.auth.kakao.service.dto.response;
-
-public record AuthenticationResponse(
-
-		String accessToken
-
-) {
-}
