+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/kakao/KakaoLoginService.java
@@ -1,11 +1,11 @@
-package com.ludo.study.studymatchingplatform.auth.kakao.service;
+package com.ludo.study.studymatchingplatform.auth.service.kakao;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.ludo.study.studymatchingplatform.auth.kakao.service.dto.KakaoOAuthToken;
-import com.ludo.study.studymatchingplatform.auth.kakao.service.dto.KakaoUserProfileDto;
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
+import com.ludo.study.studymatchingplatform.auth.service.kakao.dto.KakaoOAuthToken;
+import com.ludo.study.studymatchingplatform.auth.service.kakao.dto.KakaoUserProfileDto;
 import com.ludo.study.studymatchingplatform.study.service.exception.NotFoundException;
 import com.ludo.study.studymatchingplatform.user.domain.Social;
 import com.ludo.study.studymatchingplatform.user.domain.User;
