+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/kakao/KakaoProfileRequestService.java
@@ -1,4 +1,4 @@
-package com.ludo.study.studymatchingplatform.auth.kakao.service;
+package com.ludo.study.studymatchingplatform.auth.service.kakao;
 
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
-import com.ludo.study.studymatchingplatform.auth.kakao.service.dto.KakaoOAuthToken;
-import com.ludo.study.studymatchingplatform.auth.kakao.service.dto.KakaoUserProfileDto;
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
+import com.ludo.study.studymatchingplatform.auth.service.kakao.dto.KakaoOAuthToken;
+import com.ludo.study.studymatchingplatform.auth.service.kakao.dto.KakaoUserProfileDto;
 import com.ludo.study.studymatchingplatform.user.domain.Social;
 
 import lombok.RequiredArgsConstructor;
