+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/kakao/dto/KakaoUserPropertiesDto.java
@@ -1,12 +1,12 @@
-package com.ludo.study.studymatchingplatform.auth.kakao.service.dto;
+package com.ludo.study.studymatchingplatform.auth.service.kakao.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(value = PropertyNamingStrategies.LowerCaseStrategy.class)
-public record KakaoUserAccountDto(
+public record KakaoUserPropertiesDto(
 
-		String email
+		String nickname
 
 ) {
 }
