+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/kakao/dto/response/AuthenticationResponse.java
@@ -1,0 +1,8 @@
+package com.ludo.study.studymatchingplatform.auth.service.kakao.dto.response;
+
+public record AuthenticationResponse(
+
+		String accessToken
+
+) {
+}
